Remove redundant mpesa security trait
